Moves "jsonParse" into internal utilities
